Add files via upload
This is a personal project. The extensive analysis of the crimes committed in the state of Baltimore between January 2011 and 2016.
The analysis seeks to answer how the crimes are distributed throughout the months of the year
The most preferred weapons used by criminals
How day and night crimes differ throughout the year.
The most insecure crime neighborhoods in the top 3 districts and how they compare day and night
The weapons of choice in the top 3 crime districts

The analysis found out that there is a constant seasonal trend throughout the years, there is less crime committed in the month of February.
The distribution of crimes is normally distributed over the year. The crime peaks around the months of May and June.
However, Day crimes are at the lowest during the month of June and July, in some districts crimes committed during the night in July is almost double what is committed during the day.